Adds app/routes/broadcasts dir
