+++ app/routes/broadcasts/index.js
@@ -5,7 +5,7 @@ const express = require('express');
 const router = express.Router();
 
 // Middleware
-const indexMiddleware = require('../../lib/middleware/broadcasts-index');
+const indexMiddleware = require('../../../lib/middleware/broadcasts-index');
 
 router.get('/', indexMiddleware());
